# src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "tackle-pathfinder",
        "version": "0.0.4",
        "description": "Tackle PathFinder 0.0.4\n\nChanges : \n* added delete assessment function\n* added create assessment copying responses from another assessment\n* added bulk creation/copy of assessments\n* added identified risks function\n* added assessment confidence function\n* added multilanguage optional param for GetAssessment\n* added get questionnaire header list\n"
    },
    "paths": {
        "/assessments": {
            "summary": "Path used to manage a single assessment.",
            "description": "The REST endpoint/path used to get, update, and delete instances of an `assessment`.  \nThis path contains `GET`, `POST` operations used to perform the get a list of assessments and to create one.",
            "get": {
                "parameters": [
                    {
                        "examples": {
                            "application_null": {
                                "value": ""
                            },
                            "application_noassessed": {
                                "value": "1000"
                            },
                            "application_alreadyassessed": {
                                "value": "1"
                            }
                        },
                        "name": "applicationId",
                        "description": "",
                        "schema": {
                            "type": "integer"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AssessmentHeader"
                                    }
                                },
                                "examples": {
                                    "application_noassessed": {
                                        "value": [
                                        ]
                                    },
                                    "application_alreadyassessed": {
                                        "value": [
                                            {
                                                "status": "EMPTY",
                                                "applicationId": 68,
                                                "id": 66
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "Success"
                    },
                    "400": {
                        "content": {
                            "text/xml": {
                                "examples": {
                                    "application_null": {
                                        "value": ""
                                    }
                                }
                            }
                        },
                        "description": "When the request is not well formed."
                    }
                },
                "operationId": "getApplicationAssessments",
                "description": ""
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Application"
                            },
                            "examples": {
                                "assessment_post_ok": {
                                    "value": {
                                        "applicationId": 92
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "parameters": [
                    {
                        "examples": {
                            "post_copy_assessment": {
                                "value": "57"
                            }
                        },
                        "name": "fromAssessmentId",
                        "description": "Setting this parameter will make API to create the assessment for the applicationId specified in the body to copy all answers and comments from an already existing assessment.",
                        "schema": {
                            "type": "integer"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssessmentHeader"
                                },
                                "examples": {
                                    "assessment_post_ok": {
                                        "value": {
                                            "name": "some text",
                                            "status": "EMPTY",
                                            "applicationId": 92,
                                            "id": 3
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "description": "Invalid format"
                    },
                    "404": {
                        "description": "If the fromAssessmentId param is passed but that assessment can not be found."
                    }
                },
                "operationId": "createAssessment",
                "summary": "Creates an assessment associated to an applicationId. It can copy the responses from another assessment."
            }
        },
...
        "/assessments/bulk": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            },
                            "examples": {
                                "bulk_request": {
                                    "value": [
                                        74,
                                        54
                                    ]
                                }
                            }
                        }
                    },
                    "required": true
                },
                "parameters": [
                    {
                        "name": "fromAssessmentId",
                        "description": "",
                        "schema": {
                            "type": "integer"
                        },
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssessmentBulk"
                                },
                                "examples": {
                                    "assessment_bulk_post": {
                                        "value": {
                                            "bulkId": 58,
                                            "applications": [
                                                9,
                                                45
                                            ],
                                            "fromAssessmentId": 74,
                                            "createdAssessments": [
                                                {
                                                    "status": "COMPLETE",
                                                    "applicationId": 57,
                                                    "id": 50,
                                                    "error": "some text"
                                                },
                                                {
                                                    "status": "STARTED",
                                                    "applicationId": 6,
                                                    "id": 13,
                                                    "error": "some text"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        },
                        "description": "When the operation finishes successfully."
                    }
                },
                "operationId": "bulkCreateAssessment",
                "summary": "Async call to create a list of assessments for the given applications.",
                "description": "It will allow to copy the assessment answers from a given assessment ( in query params )"
            }
        },
        "/assessments/bulk/{bulkId}": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssessmentBulk"
                                },
                                "examples": {
                                    "bulk_response_ok": {
                                        "value": {
                                            "bulkId": 31,
                                            "applications": [
                                                92,
                                                0
                                            ],
                                            "fromAssessmentId": 72,
                                            "createdAssessments": [
                                                {
                                                    "status": "STARTED",
                                                    "applicationId": 90,
                                                    "id": 52,
                                                    "error": "some text"
                                                },
                                                {
                                                    "status": "COMPLETE",
                                                    "applicationId": 28,
                                                    "id": 18,
                                                    "error": "some text"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        },
                        "description": "If bulk process is found"
                    },
                    "404": {
                        "description": "If bulk process is not found"
                    }
                }
            },
            "parameters": [
                {
                    "name": "bulkId",
                    "schema": {
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/assessments/risks": {
            "summary": "Will retrieve the risks for filtered assessments",
            "description": "The response will contain only those applications assessed. This can end on having 0 elements in the response.",
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            },
                            "examples": {
                                "risks_ok": {
                                    "value": [
                                        {
                                            "applicationId": 90
                                        },
                                        {
                                            "applicationId": 3
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RiskLine"
                                    }
                                },
                                "examples": {
                                    "risks_response_ok": {
                                        "value": [
                                            {
                                                "category": "some text",
                                                "question": "some text",
                                                "answer": "some text",
                                                "applications": [
                                                    11,
                                                    72
                                                ]
                                            },
                                            {
                                                "category": "some text",
                                                "question": "some text",
                                                "answer": "some text",
                                                "applications": [
                                                    53,
                                                    44
                                                ]
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "Success"
                    }
                }
            }
        },
        "/assessments/confidence": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            },
                            "examples": {
                                "post_confidence_request": {
                                    "value": [
                                        {
                                            "applicationId": 5
                                        },
                                        {
                                            "applicationId": 51
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AssessmentConfidence"
                                    }
                                },
                                "examples": {
                                    "confidence_response_ok": {
                                        "value": [
                                            {
                                                "assessmentId": 48,
                                                "confidence": 83
                                            },
                                            {
                                                "assessmentId": 80,
                                                "confidence": 10
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "On success"
                    }
                }
            }
        },
        "/assessments/assessment-risk": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            },
                            "examples": {
                                "assessment-risk-ok": {
                                    "value": [
                                        {
                                            "applicationId": 5
                                        },
                                        {
                                            "applicationId": 37
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AssessmentRisk"
                                    }
                                },
                                "examples": {
                                    "assessment-risk-ok-resp": {
                                        "value": [
                                            {
                                                "assessmentId": 34,
                                                "risk": "RED"
                                            },
                                            {
                                                "assessmentId": 69,
                                                "risk": "UNKNOWN"
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "Success response"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Application": {
                "description": "",
                "required": [
                    "applicationId"
                ],
                "type": "object",
                "properties": {
                    "applicationId": {
                        "description": "",
                        "type": "integer"
                    }
                }
            },
            "AssessmentHeader": {
                "description": "",
                "required": [
                    "id",
                    "applicationId",
                    "status"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/Status",
                        "description": ""
                    },
                    "applicationId": {
                        "description": "",
                        "type": "integer"
                    },
                    "id": {
                        "description": "",
                        "type": "integer"
                    }
                },
                "example": {
                    "status": "STARTED",
                    "id": 24,
                    "applicationId": 14
                }
            },
...
            "Status": {
                "description": "Status of an Assessment",
                "enum": [
                    "STARTED",
                    "COMPLETE"
                ],
                "type": "string"
            },
...
            "AssessmentBulk": {
                "description": "",
                "required": [
                    "bulkId",
                    "applications",
                    "assessments",
                    "completed"
                ],
                "type": "object",
                "properties": {
                    "bulkId": {
                        "description": "",
                        "type": "integer"
                    },
                    "applications": {
                        "description": "",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "fromAssessmentId": {
                        "description": "",
                        "type": "integer"
                    },
                    "completed": {
                        "description": "",
                        "type": "boolean"
                    },
                    "assessments": {
                        "description": "",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AssessmentHeaderBulk"
                        }
                    }
                }
            },
            "AssessmentHeaderBulk": {
                "description": "",
                "required": [
                    "id",
                    "applicationId",
                    "status"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "$ref": "#/components/schemas/Status",
                        "description": ""
                    },
                    "applicationId": {
                        "description": "",
                        "type": "integer"
                    },
                    "id": {
                        "description": "",
                        "type": "integer"
                    },
                    "error": {
                        "description": "",
                        "type": "string"
                    }
                },
                "example": {
                    "status": "STARTED",
                    "id": 24,
                    "applicationId": 14
                }
            },
...
            "RiskLine": {
                "description": "",
                "required": [
                    "category",
                    "question",
                    "answer",
                    "applications"
                ],
                "type": "object",
                "properties": {
                    "category": {
                        "description": "",
                        "type": "string"
                    },
                    "question": {
                        "description": "",
                        "type": "string"
                    },
                    "answer": {
                        "description": "",
                        "type": "string"
                    },
                    "applications": {
                        "description": "",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                }
            },
            "AssessmentConfidence": {
                "description": "",
                "required": [
                    "assessmentId",
                    "confidence",
                    "applicationId"
                ],
                "type": "object",
                "properties": {
                    "assessmentId": {
                        "description": "",
                        "type": "integer"
                    },
                    "confidence": {
                        "description": "",
                        "type": "integer"
                    },
                    "applicationId": {
                        "description": "",
                        "type": "integer"
                    }
                }
            },
...
            "Risk": {
                "description": "",
                "enum": [
                    "RED",
                    "AMBER",
                    "GREEN",
                    "UNKNOWN"
                ],
                "type": "string"
            },
            "AssessmentRisk": {
                "description": "",
                "required": [
                    "assessmentId",
                    "risk"
                ],
                "type": "object",
                "properties": {
                    "assessmentId": {
                        "description": "",
                        "type": "integer"
                    },
                    "risk": {
                        "$ref": "#/components/schemas/Risk",
                        "description": ""
                    }
                }
            },
...
        },
...
    }
}